refactor(annotations-sample): adapt to library updates
Using annotations instead of marker interfaces for query, mutation,
subscription resolvers.
