+++ graphql-annotations/build.gradle
@@ -1,7 +1,6 @@
 apply from: "../spring-boot-commons.gradle"
 
 dependencies {
-    implementation "com.graphql-java-kickstart:graphql-spring-boot-starter:$LIB_GRAPHQL_SPRING_BOOT_VER"
     implementation "com.graphql-java-kickstart:graphql-kickstart-spring-boot-starter-graphql-annotations:$LIB_GRAPHQL_SPRING_BOOT_VER"
     runtimeOnly "com.graphql-java-kickstart:graphiql-spring-boot-starter:$LIB_GRAPHQL_SPRING_BOOT_VER"
     runtimeOnly "com.graphql-java-kickstart:playground-spring-boot-starter:$LIB_GRAPHQL_SPRING_BOOT_VER"
